# openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Naytia 像素小站 API",
  "description": "账号 / 论坛 / 工具箱收藏。契约见 docs/plans/03。",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/tools/{tool_id}": {
   "get": {
    "tags": [
     "public-v1"
    ],
    "summary": "Get Tool",
    "operationId": "get_tool_api_v1_tools__tool_id__get",
    "parameters": [
     {
      "name": "tool_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Tool Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "title": "Response Get Tool Api V1 Tools  Tool Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/tools/{tool_id}/run": {
   "post": {
    "tags": [
     "public-v1"
    ],
    "summary": "Run Tool",
    "operationId": "run_tool_api_v1_tools__tool_id__run_post",
    "parameters": [
     {
      "name": "tool_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Tool Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RunRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "title": "Response Run Tool Api V1 Tools  Tool Id  Run Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/api-keys": {
   "get": {
    "tags": [
     "public-v1-admin"
    ],
    "summary": "List Keys",
    "operationId": "list_keys_api_admin_api_keys_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "title": "Response List Keys Api Admin Api Keys Get"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "public-v1-admin"
    ],
    "summary": "Create Key",
    "operationId": "create_key_api_admin_api_keys_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ApiKeyCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "title": "Response Create Key Api Admin Api Keys Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/api-keys/{key_id}": {
   "delete": {
    "tags": [
     "public-v1-admin"
    ],
    "summary": "Disable Key",
    "operationId": "disable_key_api_admin_api_keys__key_id__delete",
    "parameters": [
     {
      "name": "key_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Key Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ApiKeyCreate": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 64,
      "minLength": 1,
      "title": "Name"
     }
    },
    "type": "object",
    "required": [
     "name"
    ],
    "title": "ApiKeyCreate"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "RunRequest": {
    "properties": {
     "input": {
      "type": "object",
      "title": "Input"
     },
     "ruleset_id": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 64
       },
       {
        "type": "null"
       }
      ],
      "title": "Ruleset Id"
     }
    },
    "type": "object",
    "title": "RunRequest"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}